core: thread: separate old SMC interface handling
Moves the old SMC interface handing parts for thread*.S into separate
files to make way for SPCI which will provide another interface towards
normal world.

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
