+++ core/arch/arm/kernel/thread.c
@@ -13,7 +13,6 @@
 #include <kernel/asan.h>
 #include <kernel/lockdep.h>
 #include <kernel/misc.h>
-#include <kernel/msg_param.h>
 #include <kernel/panic.h>
 #include <kernel/spinlock.h>
 #include <kernel/tee_ta_manager.h>
@@ -25,13 +24,8 @@
 #include <mm/tee_mm.h>
 #include <mm/tee_mmu.h>
 #include <mm/tee_pager.h>
-#include <optee_msg.h>
-#include <optee_rpc_cmd.h>
 #include <smccc.h>
-#include <sm/optee_smc.h>
 #include <sm/sm.h>
-#include <tee/tee_cryp_utl.h>
-#include <tee/tee_fs_rpc.h>
 #include <trace.h>
 #include <util.h>
 
@@ -121,7 +115,7 @@ KEEP_PAGER(stack_tmp_export);
 KEEP_PAGER(stack_tmp_stride);
 
 thread_smc_handler_t thread_std_smc_handler_ptr __nex_bss;
-static thread_smc_handler_t thread_fast_smc_handler_ptr __nex_bss;
+thread_smc_handler_t thread_fast_smc_handler_ptr __nex_bss;
 thread_nintr_handler_t thread_nintr_handler_ptr __nex_bss;
 thread_pm_handler_t thread_cpu_on_handler_ptr __nex_bss;
 thread_pm_handler_t thread_cpu_off_handler_ptr __nex_bss;
@@ -150,9 +144,6 @@ static uint8_t thread_user_kdata_page[
 #endif
 
 static unsigned int thread_global_lock __nex_bss = SPINLOCK_UNLOCK;
-static bool thread_prealloc_rpc_cache;
-
-static unsigned int thread_rpc_pnum;
 
 static void init_canaries(void)
 {
@@ -214,12 +205,12 @@ void thread_check_canaries(void)
 #endif/*CFG_WITH_STACK_CANARIES*/
 }
 
-static void lock_global(void)
+void thread_lock_global(void)
 {
 	cpu_spin_lock(&thread_global_lock);
 }
 
-static void unlock_global(void)
+void thread_unlock_global(void)
 {
 	cpu_spin_unlock(&thread_global_lock);
 }
@@ -326,8 +317,8 @@ static void thread_lazy_restore_ns_vfp(void)
 }
 
 #ifdef ARM32
-static void init_regs(struct thread_ctx *thread,
-		struct thread_smc_args *args)
+static void init_regs(struct thread_ctx *thread, uint32_t a0, uint32_t a1,
+		      uint32_t a2, uint32_t a3)
 {
 	thread->regs.pc = (uint32_t)thread_std_smc_entry;
 
@@ -348,20 +339,20 @@ static void init_regs(struct thread_ctx *thread,
 	 * Copy arguments into context. This will make the
 	 * arguments appear in r0-r7 when thread is started.
 	 */
-	thread->regs.r0 = args->a0;
-	thread->regs.r1 = args->a1;
-	thread->regs.r2 = args->a2;
-	thread->regs.r3 = args->a3;
-	thread->regs.r4 = args->a4;
-	thread->regs.r5 = args->a5;
-	thread->regs.r6 = args->a6;
-	thread->regs.r7 = args->a7;
+	thread->regs.r0 = a0;
+	thread->regs.r1 = a1;
+	thread->regs.r2 = a2;
+	thread->regs.r3 = a3;
+	thread->regs.r4 = 0;
+	thread->regs.r5 = 0;
+	thread->regs.r6 = 0;
+	thread->regs.r7 = 0;
 }
 #endif /*ARM32*/
 
 #ifdef ARM64
-static void init_regs(struct thread_ctx *thread,
-		struct thread_smc_args *args)
+static void init_regs(struct thread_ctx *thread, uint32_t a0, uint32_t a1,
+		      uint32_t a2, uint32_t a3)
 {
 	thread->regs.pc = (uint64_t)thread_std_smc_entry;
 
@@ -378,14 +369,14 @@ static void init_regs(struct thread_ctx *thread,
 	 * Copy arguments into context. This will make the
 	 * arguments appear in x0-x7 when thread is started.
 	 */
-	thread->regs.x[0] = args->a0;
-	thread->regs.x[1] = args->a1;
-	thread->regs.x[2] = args->a2;
-	thread->regs.x[3] = args->a3;
-	thread->regs.x[4] = args->a4;
-	thread->regs.x[5] = args->a5;
-	thread->regs.x[6] = args->a6;
-	thread->regs.x[7] = args->a7;
+	thread->regs.x[0] = a0;
+	thread->regs.x[1] = a1;
+	thread->regs.x[2] = a2;
+	thread->regs.x[3] = a3;
+	thread->regs.x[4] = 0;
+	thread->regs.x[5] = 0;
+	thread->regs.x[6] = 0;
+	thread->regs.x[7] = 0;
 
 	/* Set up frame pointer as per the Aarch64 AAPCS */
 	thread->regs.x[29] = 0;
@@ -412,7 +403,7 @@ void thread_clr_boot_thread(void)
 	l->curr_thread = -1;
 }
 
-static void thread_alloc_and_run(struct thread_smc_args *args)
+void thread_alloc_and_run(uint32_t a0, uint32_t a1, uint32_t a2, uint32_t a3)
 {
 	size_t n;
 	struct thread_core_local *l = thread_get_core_local();
@@ -420,7 +411,7 @@ static void thread_alloc_and_run(struct thread_smc_args *args)
 
 	assert(l->curr_thread == -1);
 
-	lock_global();
+	thread_lock_global();
 
 	for (n = 0; n < CFG_NUM_THREADS; n++) {
 		if (threads[n].state == THREAD_STATE_FREE) {
@@ -430,56 +421,55 @@ static void thread_alloc_and_run(struct thread_smc_args *args)
 		}
 	}
 
-	unlock_global();
+	thread_unlock_global();
 
-	if (!found_thread) {
-		args->a0 = OPTEE_SMC_RETURN_ETHREAD_LIMIT;
+	if (!found_thread)
 		return;
-	}
 
 	l->curr_thread = n;
 
 	threads[n].flags = 0;
-	init_regs(threads + n, args);
-
-	/* Save Hypervisor Client ID */
-	threads[n].hyp_clnt_id = args->a7;
+	init_regs(threads + n, a0, a1, a2, a3);
 
 	thread_lazy_save_ns_vfp();
 	thread_resume(&threads[n].regs);
+	/*NOTREACHED*/
+	panic();
 }
 
 #ifdef ARM32
-static void copy_a0_to_a5(struct thread_ctx_regs *regs,
-		struct thread_smc_args *args)
+static void copy_a0_to_a5(struct thread_ctx_regs *regs, uint32_t a0,
+			  uint32_t a1, uint32_t a2, uint32_t a3, uint32_t a4,
+			  uint32_t a5)
 {
 	/*
 	 * Update returned values from RPC, values will appear in
 	 * r0-r3 when thread is resumed.
 	 */
-	regs->r0 = args->a0;
-	regs->r1 = args->a1;
-	regs->r2 = args->a2;
-	regs->r3 = args->a3;
-	regs->r4 = args->a4;
-	regs->r5 = args->a5;
+	regs->r0 = a0;
+	regs->r1 = a1;
+	regs->r2 = a2;
+	regs->r3 = a3;
+	regs->r4 = a4;
+	regs->r5 = a5;
 }
 #endif /*ARM32*/
 
 #ifdef ARM64
-static void copy_a0_to_a5(struct thread_ctx_regs *regs,
-		struct thread_smc_args *args)
+static void copy_a0_to_a5(struct thread_ctx_regs *regs, uint32_t a0,
+			  uint32_t a1, uint32_t a2, uint32_t a3, uint32_t a4,
+			  uint32_t a5)
 {
 	/*
 	 * Update returned values from RPC, values will appear in
 	 * x0-x3 when thread is resumed.
 	 */
-	regs->x[0] = args->a0;
-	regs->x[1] = args->a1;
-	regs->x[2] = args->a2;
-	regs->x[3] = args->a3;
-	regs->x[4] = args->a4;
-	regs->x[5] = args->a5;
+	regs->x[0] = a0;
+	regs->x[1] = a1;
+	regs->x[2] = a2;
+	regs->x[3] = a3;
+	regs->x[4] = a4;
+	regs->x[5] = a5;
 }
 #endif /*ARM64*/
 
@@ -507,29 +497,26 @@ static bool is_user_mode(struct thread_ctx_regs *regs)
 	return is_from_user((uint32_t)regs->cpsr);
 }
 
-static void thread_resume_from_rpc(struct thread_smc_args *args)
+void thread_resume_from_rpc(uint32_t a0, uint32_t a1, uint32_t a2, uint32_t a3,
+			    uint32_t a4, uint32_t a5)
 {
-	size_t n = args->a3; /* thread id */
+	size_t n = a3; /* thread id */
 	struct thread_core_local *l = thread_get_core_local();
-	uint32_t rv = 0;
+	bool found_thread = false;
 
 	assert(l->curr_thread == -1);
 
-	lock_global();
+	thread_lock_global();
 
-	if (n < CFG_NUM_THREADS &&
-	    threads[n].state == THREAD_STATE_SUSPENDED &&
-	    args->a7 == threads[n].hyp_clnt_id)
+	if (n < CFG_NUM_THREADS && threads[n].state == THREAD_STATE_SUSPENDED) {
 		threads[n].state = THREAD_STATE_ACTIVE;
-	else
-		rv = OPTEE_SMC_RETURN_ERESUME;
-
-	unlock_global();
-
-	if (rv) {
-		args->a0 = rv;
-		return;
+		found_thread = true;
 	}
+
+	thread_unlock_global();
+
+	if (!found_thread)
+		return;
 
 	l->curr_thread = n;
 
@@ -544,99 +531,14 @@ static void thread_resume_from_rpc(struct thread_smc_args *args)
 	 * get parameters from non-secure world.
 	 */
 	if (threads[n].flags & THREAD_FLAGS_COPY_ARGS_ON_RETURN) {
-		copy_a0_to_a5(&threads[n].regs, args);
+		copy_a0_to_a5(&threads[n].regs, a0, a1, a2, a3, a4, a5);
 		threads[n].flags &= ~THREAD_FLAGS_COPY_ARGS_ON_RETURN;
 	}
 
 	thread_lazy_save_ns_vfp();
 	thread_resume(&threads[n].regs);
-}
-
-void thread_handle_fast_smc(struct thread_smc_args *args)
-{
-	thread_check_canaries();
-
-#ifdef CFG_VIRTUALIZATION
-	if (!virt_set_guest(args->a7)) {
-		args->a0 = OPTEE_SMC_RETURN_ENOTAVAIL;
-		goto out;
-	}
-#endif
-
-	thread_fast_smc_handler_ptr(args);
-
-#ifdef CFG_VIRTUALIZATION
-	virt_unset_guest();
-#endif
-	/* Fast handlers must not unmask any exceptions */
-out:
-	__maybe_unused;
-	assert(thread_get_exceptions() == THREAD_EXCP_ALL);
-}
-
-void thread_handle_std_smc(struct thread_smc_args *args)
-{
-	thread_check_canaries();
-
-#ifdef CFG_VIRTUALIZATION
-	if (!virt_set_guest(args->a7)) {
-		args->a0 = OPTEE_SMC_RETURN_ENOTAVAIL;
-		return;
-	}
-#endif
-
-	if (args->a0 == OPTEE_SMC_CALL_RETURN_FROM_RPC)
-		thread_resume_from_rpc(args);
-	else
-		thread_alloc_and_run(args);
-
-#ifdef CFG_VIRTUALIZATION
-	virt_unset_guest();
-#endif
-
-}
-
-/**
- * Free physical memory previously allocated with thread_rpc_alloc_arg()
- *
- * @cookie:	cookie received when allocating the buffer
- */
-static void thread_rpc_free_arg(uint64_t cookie)
-{
-	if (cookie) {
-		uint32_t rpc_args[THREAD_RPC_NUM_ARGS] = {
-			OPTEE_SMC_RETURN_RPC_FREE
-		};
-
-		reg_pair_from_64(cookie, rpc_args + 1, rpc_args + 2);
-		thread_rpc(rpc_args);
-	}
-}
-
-/*
- * Helper routine for the assembly function thread_std_smc_entry()
- *
- * Note: this function is weak just to make it possible to exclude it from
- * the unpaged area.
- */
-void __weak __thread_std_smc_entry(struct thread_smc_args *args)
-{
-#ifdef CFG_VIRTUALIZATION
-	virt_on_stdcall();
-#endif
-	thread_std_smc_handler_ptr(args);
-
-	if (args->a0 == OPTEE_SMC_RETURN_OK) {
-		struct thread_ctx *thr = threads + thread_get_id();
-
-		tee_fs_rpc_cache_clear(&thr->tsd);
-		if (!thread_prealloc_rpc_cache) {
-			thread_rpc_free_arg(mobj_get_cookie(thr->rpc_mobj));
-			mobj_free(thr->rpc_mobj);
-			thr->rpc_arg = 0;
-			thr->rpc_mobj = NULL;
-		}
-	}
+	/*NOTREACHED*/
+	panic();
 }
 
 void *thread_get_tmp_sp(void)
@@ -715,7 +617,7 @@ void thread_state_free(void)
 		(void *)(threads[ct].stack_va_end - STACK_THREAD_SIZE),
 		STACK_THREAD_SIZE);
 
-	lock_global();
+	thread_lock_global();
 
 	assert(threads[ct].state == THREAD_STATE_ACTIVE);
 	threads[ct].state = THREAD_STATE_FREE;
@@ -725,7 +627,7 @@ void thread_state_free(void)
 #ifdef CFG_VIRTUALIZATION
 	virt_unset_guest();
 #endif
-	unlock_global();
+	thread_unlock_global();
 }
 
 #ifdef CFG_WITH_PAGER
@@ -773,7 +675,7 @@ int thread_state_suspend(uint32_t flags, uint32_t cpsr, vaddr_t pc)
 	}
 	thread_lazy_restore_ns_vfp();
 
-	lock_global();
+	thread_lock_global();
 
 	assert(threads[ct].state == THREAD_STATE_ACTIVE);
 	threads[ct].flags |= flags;
@@ -793,7 +695,7 @@ int thread_state_suspend(uint32_t flags, uint32_t cpsr, vaddr_t pc)
 	virt_unset_guest();
 #endif
 
-	unlock_global();
+	thread_unlock_global();
 
 	return ct;
 }
@@ -1344,384 +1246,3 @@ void thread_get_user_kdata(struct mobj **mobj, size_t *offset,
 	*sz = sizeof(thread_user_kdata_page);
 }
 #endif
-
-bool thread_disable_prealloc_rpc_cache(uint64_t *cookie)
-{
-	bool rv;
-	size_t n;
-	uint32_t exceptions = thread_mask_exceptions(THREAD_EXCP_FOREIGN_INTR);
-
-	lock_global();
-
-	for (n = 0; n < CFG_NUM_THREADS; n++) {
-		if (threads[n].state != THREAD_STATE_FREE) {
-			rv = false;
-			goto out;
-		}
-	}
-
-	rv = true;
-	for (n = 0; n < CFG_NUM_THREADS; n++) {
-		if (threads[n].rpc_arg) {
-			*cookie = mobj_get_cookie(threads[n].rpc_mobj);
-			mobj_free(threads[n].rpc_mobj);
-			threads[n].rpc_arg = NULL;
-			goto out;
-		}
-	}
-
-	*cookie = 0;
-	thread_prealloc_rpc_cache = false;
-out:
-	unlock_global();
-	thread_unmask_exceptions(exceptions);
-	return rv;
-}
-
-bool thread_enable_prealloc_rpc_cache(void)
-{
-	bool rv;
-	size_t n;
-	uint32_t exceptions = thread_mask_exceptions(THREAD_EXCP_FOREIGN_INTR);
-
-	lock_global();
-
-	for (n = 0; n < CFG_NUM_THREADS; n++) {
-		if (threads[n].state != THREAD_STATE_FREE) {
-			rv = false;
-			goto out;
-		}
-	}
-
-	rv = true;
-	thread_prealloc_rpc_cache = true;
-out:
-	unlock_global();
-	thread_unmask_exceptions(exceptions);
-	return rv;
-}
-
-/**
- * Allocates data for struct optee_msg_arg.
- *
- * @size:	size in bytes of struct optee_msg_arg
- *
- * @returns	mobj that describes allocated buffer or NULL on error
- */
-static struct mobj *thread_rpc_alloc_arg(size_t size)
-{
-	paddr_t pa;
-	uint64_t co;
-	uint32_t rpc_args[THREAD_RPC_NUM_ARGS] = {
-		OPTEE_SMC_RETURN_RPC_ALLOC, size
-	};
-	struct mobj *mobj = NULL;
-
-	thread_rpc(rpc_args);
-
-	pa = reg_pair_to_64(rpc_args[1], rpc_args[2]);
-	co = reg_pair_to_64(rpc_args[4], rpc_args[5]);
-
-	if (!ALIGNMENT_IS_OK(pa, struct optee_msg_arg))
-		goto err;
-
-	/* Check if this region is in static shared space */
-	if (core_pbuf_is(CORE_MEM_NSEC_SHM, pa, size))
-		mobj = mobj_shm_alloc(pa, size, co);
-#ifdef CFG_CORE_DYN_SHM
-	else if ((!(pa & SMALL_PAGE_MASK)) && size <= SMALL_PAGE_SIZE)
-		mobj = mobj_mapped_shm_alloc(&pa, 1, 0, co);
-#endif
-
-	if (!mobj)
-		goto err;
-
-	return mobj;
-err:
-	thread_rpc_free_arg(co);
-	mobj_free(mobj);
-	return NULL;
-}
-
-static bool set_rmem(struct optee_msg_param *param,
-		     struct thread_param *tpm)
-{
-	param->attr = tpm->attr - THREAD_PARAM_ATTR_MEMREF_IN +
-		      OPTEE_MSG_ATTR_TYPE_RMEM_INPUT;
-	param->u.rmem.offs = tpm->u.memref.offs;
-	param->u.rmem.size = tpm->u.memref.size;
-	if (tpm->u.memref.mobj) {
-		param->u.rmem.shm_ref = mobj_get_cookie(tpm->u.memref.mobj);
-		if (!param->u.rmem.shm_ref)
-			return false;
-	} else {
-		param->u.rmem.shm_ref = 0;
-	}
-
-	return true;
-}
-
-static bool set_tmem(struct optee_msg_param *param,
-		     struct thread_param *tpm)
-{
-	paddr_t pa = 0;
-	uint64_t shm_ref = 0;
-	struct mobj *mobj = tpm->u.memref.mobj;
-
-	param->attr = tpm->attr - THREAD_PARAM_ATTR_MEMREF_IN +
-		      OPTEE_MSG_ATTR_TYPE_TMEM_INPUT;
-	if (mobj) {
-		shm_ref = mobj_get_cookie(mobj);
-		if (!shm_ref)
-			return false;
-		if (mobj_get_pa(mobj, tpm->u.memref.offs, 0, &pa))
-			return false;
-	}
-
-	param->u.tmem.size = tpm->u.memref.size;
-	param->u.tmem.buf_ptr = pa;
-	param->u.tmem.shm_ref = shm_ref;
-
-	return true;
-}
-
-static uint32_t get_rpc_arg(uint32_t cmd, size_t num_params,
-			    struct thread_param *params, void **arg_ret,
-			    uint64_t *carg_ret)
-{
-	struct thread_ctx *thr = threads + thread_get_id();
-	struct optee_msg_arg *arg = thr->rpc_arg;
-	size_t sz = OPTEE_MSG_GET_ARG_SIZE(THREAD_RPC_MAX_NUM_PARAMS);
-
-	if (num_params > THREAD_RPC_MAX_NUM_PARAMS)
-		return TEE_ERROR_BAD_PARAMETERS;
-
-	if (!arg) {
-		struct mobj *mobj = thread_rpc_alloc_arg(sz);
-
-		if (!mobj)
-			return TEE_ERROR_OUT_OF_MEMORY;
-
-		arg = mobj_get_va(mobj, 0);
-		if (!arg) {
-			thread_rpc_free_arg(mobj_get_cookie(mobj));
-			return TEE_ERROR_OUT_OF_MEMORY;
-		}
-
-		thr->rpc_arg = arg;
-		thr->rpc_mobj = mobj;
-	}
-
-	memset(arg, 0, OPTEE_MSG_GET_ARG_SIZE(num_params));
-	arg->cmd = cmd;
-	arg->num_params = num_params;
-	arg->ret = TEE_ERROR_GENERIC; /* in case value isn't updated */
-
-	for (size_t n = 0; n < num_params; n++) {
-		switch (params[n].attr) {
-		case THREAD_PARAM_ATTR_NONE:
-			arg->params[n].attr = OPTEE_MSG_ATTR_TYPE_NONE;
-			break;
-		case THREAD_PARAM_ATTR_VALUE_IN:
-		case THREAD_PARAM_ATTR_VALUE_OUT:
-		case THREAD_PARAM_ATTR_VALUE_INOUT:
-			arg->params[n].attr = params[n].attr -
-					      THREAD_PARAM_ATTR_VALUE_IN +
-					      OPTEE_MSG_ATTR_TYPE_VALUE_INPUT;
-			arg->params[n].u.value.a = params[n].u.value.a;
-			arg->params[n].u.value.b = params[n].u.value.b;
-			arg->params[n].u.value.c = params[n].u.value.c;
-			break;
-		case THREAD_PARAM_ATTR_MEMREF_IN:
-		case THREAD_PARAM_ATTR_MEMREF_OUT:
-		case THREAD_PARAM_ATTR_MEMREF_INOUT:
-			if (!params[n].u.memref.mobj ||
-			    mobj_matches(params[n].u.memref.mobj,
-					 CORE_MEM_NSEC_SHM)) {
-				if (!set_tmem(arg->params + n, params + n))
-					return TEE_ERROR_BAD_PARAMETERS;
-			} else  if (mobj_matches(params[n].u.memref.mobj,
-						 CORE_MEM_REG_SHM)) {
-				if (!set_rmem(arg->params + n, params + n))
-					return TEE_ERROR_BAD_PARAMETERS;
-			} else {
-				return TEE_ERROR_BAD_PARAMETERS;
-			}
-			break;
-		default:
-			return TEE_ERROR_BAD_PARAMETERS;
-		}
-	}
-
-	*arg_ret = arg;
-	*carg_ret = mobj_get_cookie(thr->rpc_mobj);
-
-	return TEE_SUCCESS;
-}
-
-static uint32_t get_rpc_arg_res(struct optee_msg_arg *arg, size_t num_params,
-				struct thread_param *params)
-{
-	for (size_t n = 0; n < num_params; n++) {
-		switch (params[n].attr) {
-		case THREAD_PARAM_ATTR_VALUE_OUT:
-		case THREAD_PARAM_ATTR_VALUE_INOUT:
-			params[n].u.value.a = arg->params[n].u.value.a;
-			params[n].u.value.b = arg->params[n].u.value.b;
-			params[n].u.value.c = arg->params[n].u.value.c;
-			break;
-		case THREAD_PARAM_ATTR_MEMREF_OUT:
-		case THREAD_PARAM_ATTR_MEMREF_INOUT:
-			/*
-			 * rmem.size and tmem.size is the same type and
-			 * location.
-			 */
-			params[n].u.memref.size = arg->params[n].u.rmem.size;
-			break;
-		default:
-			break;
-		}
-	}
-
-	return arg->ret;
-}
-
-uint32_t thread_rpc_cmd(uint32_t cmd, size_t num_params,
-			struct thread_param *params)
-{
-	uint32_t rpc_args[THREAD_RPC_NUM_ARGS] = { OPTEE_SMC_RETURN_RPC_CMD };
-	void *arg = NULL;
-	uint64_t carg = 0;
-	uint32_t ret = 0;
-
-	/* The source CRYPTO_RNG_SRC_JITTER_RPC is safe to use here */
-	plat_prng_add_jitter_entropy(CRYPTO_RNG_SRC_JITTER_RPC,
-				     &thread_rpc_pnum);
-
-	ret = get_rpc_arg(cmd, num_params, params, &arg, &carg);
-	if (ret)
-		return ret;
-
-	reg_pair_from_64(carg, rpc_args + 1, rpc_args + 2);
-	thread_rpc(rpc_args);
-
-	return get_rpc_arg_res(arg, num_params, params);
-}
-
-/**
- * Free physical memory previously allocated with thread_rpc_alloc()
- *
- * @cookie:	cookie received when allocating the buffer
- * @bt:		must be the same as supplied when allocating
- * @mobj:	mobj that describes allocated buffer
- *
- * This function also frees corresponding mobj.
- */
-static void thread_rpc_free(unsigned int bt, uint64_t cookie, struct mobj *mobj)
-{
-	uint32_t rpc_args[THREAD_RPC_NUM_ARGS] = { OPTEE_SMC_RETURN_RPC_CMD };
-	void *arg = NULL;
-	uint64_t carg = 0;
-	struct thread_param param = THREAD_PARAM_VALUE(IN, bt, cookie, 0);
-	uint32_t ret = get_rpc_arg(OPTEE_RPC_CMD_SHM_FREE, 1, &param,
-				   &arg, &carg);
-
-	mobj_free(mobj);
-
-	if (!ret) {
-		reg_pair_from_64(carg, rpc_args + 1, rpc_args + 2);
-		thread_rpc(rpc_args);
-	}
-}
-
-static struct mobj *get_rpc_alloc_res(struct optee_msg_arg *arg,
-				      unsigned int bt, size_t size)
-{
-	struct mobj *mobj = NULL;
-	uint64_t cookie = 0;
-	size_t psize = 0;
-	uint64_t attr = 0;
-
-	if (arg->ret || arg->num_params != 1)
-		return NULL;
-
-	psize = READ_ONCE(arg->params[0].u.tmem.size);
-	if (psize < size)
-		return NULL;
-
-	attr = READ_ONCE(arg->params[0].attr);
-	if (attr == OPTEE_MSG_ATTR_TYPE_TMEM_OUTPUT) {
-		cookie = arg->params[0].u.tmem.shm_ref;
-		mobj = mobj_shm_alloc(arg->params[0].u.tmem.buf_ptr,
-				      psize,
-				      cookie);
-	} else if (attr == (OPTEE_MSG_ATTR_TYPE_TMEM_OUTPUT |
-			    OPTEE_MSG_ATTR_NONCONTIG)) {
-		cookie = arg->params[0].u.tmem.shm_ref;
-		mobj = msg_param_mobj_from_noncontig(
-			arg->params[0].u.tmem.buf_ptr,
-			psize,
-			cookie,
-			true);
-	} else {
-		return NULL;
-	}
-
-	if (!mobj) {
-		thread_rpc_free(bt, cookie, mobj);
-		return NULL;
-	}
-
-	assert(mobj_is_nonsec(mobj));
-
-	return mobj;
-}
-
-/**
- * Allocates shared memory buffer via RPC
- *
- * @size:	size in bytes of shared memory buffer
- * @align:	required alignment of buffer
- * @bt:		buffer type OPTEE_RPC_SHM_TYPE_*
- *
- * Returns a pointer to MOBJ for the memory on success, or NULL on failure.
- */
-static struct mobj *thread_rpc_alloc(size_t size, size_t align, unsigned int bt)
-{
-	uint32_t rpc_args[THREAD_RPC_NUM_ARGS] = { OPTEE_SMC_RETURN_RPC_CMD };
-	void *arg = NULL;
-	uint64_t carg = 0;
-	struct thread_param param = THREAD_PARAM_VALUE(IN, bt, size, align);
-	uint32_t ret = get_rpc_arg(OPTEE_RPC_CMD_SHM_ALLOC, 1, &param,
-				   &arg, &carg);
-
-	if (ret)
-		return NULL;
-
-	reg_pair_from_64(carg, rpc_args + 1, rpc_args + 2);
-	thread_rpc(rpc_args);
-
-	return get_rpc_alloc_res(arg, bt, size);
-}
-
-struct mobj *thread_rpc_alloc_payload(size_t size)
-{
-	return thread_rpc_alloc(size, 8, OPTEE_RPC_SHM_TYPE_APPL);
-}
-
-void thread_rpc_free_payload(struct mobj *mobj)
-{
-	thread_rpc_free(OPTEE_RPC_SHM_TYPE_APPL, mobj_get_cookie(mobj),
-			mobj);
-}
-
-struct mobj *thread_rpc_alloc_global_payload(size_t size)
-{
-	return thread_rpc_alloc(size, 8, OPTEE_RPC_SHM_TYPE_GLOBAL);
-}
-
-void thread_rpc_free_global_payload(struct mobj *mobj)
-{
-	thread_rpc_free(OPTEE_RPC_SHM_TYPE_GLOBAL, mobj_get_cookie(mobj),
-			mobj);
-}
