+++ core/arch/arm/kernel/thread_a32.S
@@ -14,9 +14,6 @@
 #include <kernel/thread_defs.h>
 #include <kernel/unwind.h>
 #include <mm/core_mmu.h>
-#include <sm/optee_smc.h>
-#include <sm/teesmc_opteed.h>
-#include <sm/teesmc_opteed_macros.h>
 
 #include "thread_private.h"
 
@@ -30,155 +27,6 @@
 		 */
 		tst	\reg, #0x0f
 	.endm
-
-LOCAL_FUNC vector_std_smc_entry , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
-	push	{r0-r7}
-	mov	r0, sp
-	bl	thread_handle_std_smc
-	/*
-	 * Normally thread_handle_std_smc() should return via
-	 * thread_exit(), thread_rpc(), but if thread_handle_std_smc()
-	 * hasn't switched stack (error detected) it will do a normal "C"
-	 * return.
-	 */
-	pop	{r1-r8}
-	ldr	r0, =TEESMC_OPTEED_RETURN_CALL_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-UNWIND(	.fnend)
-END_FUNC vector_std_smc_entry
-
-LOCAL_FUNC vector_fast_smc_entry , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
-	push	{r0-r7}
-	mov	r0, sp
-	bl	thread_handle_fast_smc
-	pop	{r1-r8}
-	ldr	r0, =TEESMC_OPTEED_RETURN_CALL_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-UNWIND(	.fnend)
-END_FUNC vector_fast_smc_entry
-
-LOCAL_FUNC vector_fiq_entry , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
- 	/* Secure Monitor received a FIQ and passed control to us. */
-	bl	thread_check_canaries
-	ldr	lr, =thread_nintr_handler_ptr
- 	ldr	lr, [lr]
- 	blx	lr
-	mov	r1, r0
-	ldr	r0, =TEESMC_OPTEED_RETURN_FIQ_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-UNWIND(	.fnend)
-END_FUNC vector_fiq_entry
-
-LOCAL_FUNC vector_cpu_on_entry , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
-	ldr	lr, =thread_cpu_on_handler_ptr
-	ldr	lr, [lr]
-	blx	lr
-	mov	r1, r0
-	ldr	r0, =TEESMC_OPTEED_RETURN_ON_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-UNWIND(	.fnend)
-END_FUNC vector_cpu_on_entry
-
-LOCAL_FUNC vector_cpu_off_entry , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
-	ldr	lr, =thread_cpu_off_handler_ptr
-	ldr	lr, [lr]
-	blx	lr
-	mov	r1, r0
-	ldr	r0, =TEESMC_OPTEED_RETURN_OFF_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-UNWIND(	.fnend)
-END_FUNC vector_cpu_off_entry
-
-LOCAL_FUNC vector_cpu_suspend_entry , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
-	ldr	lr, =thread_cpu_suspend_handler_ptr
-	ldr	lr, [lr]
-	blx	lr
-	mov	r1, r0
-	ldr	r0, =TEESMC_OPTEED_RETURN_SUSPEND_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-UNWIND(	.fnend)
-END_FUNC vector_cpu_suspend_entry
-
-LOCAL_FUNC vector_cpu_resume_entry , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
-	ldr	lr, =thread_cpu_resume_handler_ptr
-	ldr	lr, [lr]
-	blx	lr
-	mov	r1, r0
-	ldr	r0, =TEESMC_OPTEED_RETURN_RESUME_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-UNWIND(	.fnend)
-END_FUNC vector_cpu_resume_entry
-
-LOCAL_FUNC vector_system_off_entry , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
-	ldr	lr, =thread_system_off_handler_ptr
-	ldr	lr, [lr]
-	blx	lr
-	mov	r1, r0
-	ldr	r0, =TEESMC_OPTEED_RETURN_SYSTEM_OFF_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-UNWIND(	.fnend)
-END_FUNC vector_system_off_entry
-
-LOCAL_FUNC vector_system_reset_entry , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
-	ldr	lr, =thread_system_reset_handler_ptr
-	ldr	lr, [lr]
-	blx	lr
-	mov	r1, r0
-	ldr	r0, =TEESMC_OPTEED_RETURN_SYSTEM_RESET_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-UNWIND(	.fnend)
-END_FUNC vector_system_reset_entry
-
-/*
- * Vector table supplied to ARM Trusted Firmware (ARM-TF) at
- * initialization.  Also used when compiled with the internal monitor, but
- * the cpu_*_entry and system_*_entry are not used then.
- *
- * Note that ARM-TF depends on the layout of this vector table, any change
- * in layout has to be synced with ARM-TF.
- */
-FUNC thread_vector_table , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
-	b	vector_std_smc_entry
-	b	vector_fast_smc_entry
-	b	vector_cpu_on_entry
-	b	vector_cpu_off_entry
-	b	vector_cpu_resume_entry
-	b	vector_cpu_suspend_entry
-	b	vector_fiq_entry
-	b	vector_system_off_entry
-	b	vector_system_reset_entry
-UNWIND(	.fnend)
-END_FUNC thread_vector_table
-KEEP_PAGER thread_vector_table
 
 FUNC thread_set_abt_sp , :
 UNWIND(	.fnstart)
@@ -311,7 +159,7 @@ END_FUNC thread_save_state_fiq
  * Disables IRQ and FIQ and saves state of thread, returns original
  * CPSR.
  */
-LOCAL_FUNC thread_save_state , :
+FUNC thread_save_state , :
 UNWIND(	.fnstart)
 UNWIND(	.cantunwind)
 	push	{r12, lr}
@@ -358,94 +206,6 @@ UNWIND(	.cantunwind)
 	bx	lr
 UNWIND(	.fnend)
 END_FUNC thread_save_state
-
-FUNC thread_std_smc_entry , :
-UNWIND(	.fnstart)
-UNWIND(	.cantunwind)
-	/* Pass r0-r7 in a struct thread_smc_args */
-	push	{r0-r7}
-	mov	r0, sp
-	bl	__thread_std_smc_entry
-	/*
-	 * Load the returned r0-r3 into preserved registers and skip the
-	 * "returned" r4-r7 since they will not be returned to normal
-	 * world.
-	 */
-	pop	{r4-r7}
-	add	sp, #(4 * 4)
-
-	/* Disable interrupts before switching to temporary stack */
-	cpsid	aif
-	bl	thread_get_tmp_sp
-	mov	sp, r0
-
-	bl	thread_state_free
-
-	ldr	r0, =TEESMC_OPTEED_RETURN_CALL_DONE
-	mov	r1, r4
-	mov	r2, r5
-	mov	r3, r6
-	mov	r4, r7
-	smc	#0
-	b	.	/* SMC should not return */
-UNWIND(	.fnend)
-END_FUNC thread_std_smc_entry
-
-
-/* void thread_rpc(uint32_t rv[THREAD_RPC_NUM_ARGS]) */
-FUNC thread_rpc , :
-/*
- * r0-r2 are used to pass parameters to normal world
- * r0-r5 are used to pass return vaule back from normal world
- *
- * note that r3 is used to pass "resume information", that is, which
- * thread it is that should resume.
- *
- * Since the this function is following AAPCS we need to preserve r4-r5
- * which are otherwise modified when returning back from normal world.
- */
-UNWIND(	.fnstart)
-	push	{r4-r5, lr}
-UNWIND(	.save	{r4-r5, lr})
-	push	{r0}
-UNWIND(	.save	{r0})
-
-	bl	thread_save_state
-	mov	r4, r0			/* Save original CPSR */
-
-	/*
- 	 * Switch to temporary stack and SVC mode. Save CPSR to resume into.
-	 */
-	bl	thread_get_tmp_sp
-	ldr	r5, [sp]		/* Get pointer to rv[] */
-	cps	#CPSR_MODE_SVC		/* Change to SVC mode */
-	mov	sp, r0			/* Switch to tmp stack */
-
-	mov	r0, #THREAD_FLAGS_COPY_ARGS_ON_RETURN
-	mov	r1, r4			/* CPSR to restore */
-	ldr	r2, =.thread_rpc_return
-	bl	thread_state_suspend
-	mov	r4, r0			/* Supply thread index */
-	ldr	r0, =TEESMC_OPTEED_RETURN_CALL_DONE
-	ldm	r5, {r1-r3}		/* Load rv[] into r0-r2 */
-	smc	#0
-	b	.	/* SMC should not return */
-
-.thread_rpc_return:
-	/*
-	 * At this point has the stack pointer been restored to the value
-	 * it had when thread_save_state() was called above.
-	 *
-	 * Jumps here from thread_resume above when RPC has returned. The
-	 * IRQ and FIQ bits are restored to what they where when this
-	 * function was originally entered.
-	 */
-	pop	{r12}			/* Get pointer to rv[] */
-	stm	r12, {r0-r5}		/* Store r0-r5 into rv[] */
-	pop	{r4-r5, pc}
-UNWIND(	.fnend)
-END_FUNC thread_rpc
-KEEP_PAGER thread_rpc
 
 /*
  * unsigned long thread_smc(unsigned long func_id, unsigned long a1,
@@ -737,21 +497,6 @@ END_FUNC thread_unwind_user_mode
 	bl	thread_save_state
 	.endif
 
-	mov	r0, #THREAD_FLAGS_EXIT_ON_FOREIGN_INTR
-	mrs	r1, spsr
-	pop	{r2}
-	pop	{r12}
-	blx	thread_state_suspend
-	mov	r4, r0		/* Supply thread index */
-
-	/*
-	 * Switch to SVC mode and copy current stack pointer as it already
-	 * is the tmp stack.
-	 */
-	mov	r0, sp
-	cps	#CPSR_MODE_SVC
-	mov	sp, r0
-
 #ifdef CFG_CORE_WORKAROUND_NSITR_CACHE_PRIME
 	/*
 	 * Prevent leaking information about which entries has been used in
@@ -763,13 +508,22 @@ END_FUNC thread_unwind_user_mode
 	write_iciallu
 #endif
 
-	ldr	r0, =TEESMC_OPTEED_RETURN_CALL_DONE
-	ldr	r1, =OPTEE_SMC_RETURN_RPC_FOREIGN_INTR
-	mov	r2, #0
-	mov	r3, #0
-	/* r4 is already filled in above */
-	smc	#0
-	b	.	/* SMC should not return */
+	mov	r0, #THREAD_FLAGS_EXIT_ON_FOREIGN_INTR
+	mrs	r1, spsr
+	pop	{r2}
+	pop	{r12}
+	blx	thread_state_suspend
+
+	/*
+	 * Switch to SVC mode and copy current stack pointer as it already
+	 * is the tmp stack.
+	 */
+	mov	r1, sp
+	cps	#CPSR_MODE_SVC
+	mov	sp, r1
+
+	/* Passing thread index in r0 */
+	b	thread_foreign_intr_exit
 .endm
 
 	.section .text.thread_excp_vect
