+++ core/arch/arm/kernel/thread_a64.S
@@ -12,9 +12,6 @@
 #include <kernel/thread_defs.h>
 #include <mm/core_mmu.h>
 #include <smccc.h>
-#include <sm/optee_smc.h>
-#include <sm/teesmc_opteed.h>
-#include <sm/teesmc_opteed_macros.h>
 
 #include "thread_private.h"
 
@@ -31,128 +28,6 @@
 		tst	\reg, #(SPSR_64_MODE_EL_MASK << SPSR_64_MODE_EL_SHIFT)
 		b.eq	\label
 	.endm
-
-LOCAL_FUNC vector_std_smc_entry , :
-	sub	sp, sp, #THREAD_SMC_ARGS_SIZE
-	store_xregs sp, THREAD_SMC_ARGS_X0, 0, 7
-	mov	x0, sp
-	bl	thread_handle_std_smc
-	/*
-	 * Normally thread_handle_std_smc() should return via
-	 * thread_exit(), thread_rpc(), but if thread_handle_std_smc()
-	 * hasn't switched stack (error detected) it will do a normal "C"
-	 * return.
-	 */
-	load_xregs sp, THREAD_SMC_ARGS_X0, 1, 8
-	add	sp, sp, #THREAD_SMC_ARGS_SIZE
-	ldr	x0, =TEESMC_OPTEED_RETURN_CALL_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-END_FUNC vector_std_smc_entry
-
-LOCAL_FUNC vector_fast_smc_entry , :
-	sub	sp, sp, #THREAD_SMC_ARGS_SIZE
-	store_xregs sp, THREAD_SMC_ARGS_X0, 0, 7
-	mov	x0, sp
-	bl	thread_handle_fast_smc
-	load_xregs sp, THREAD_SMC_ARGS_X0, 1, 8
-	add	sp, sp, #THREAD_SMC_ARGS_SIZE
-	ldr	x0, =TEESMC_OPTEED_RETURN_CALL_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-END_FUNC vector_fast_smc_entry
-
-LOCAL_FUNC vector_fiq_entry , :
-	/* Secure Monitor received a FIQ and passed control to us. */
-	bl	thread_check_canaries
-	adr	x16, thread_nintr_handler_ptr
-	ldr	x16, [x16]
-	blr	x16
-	ldr	x0, =TEESMC_OPTEED_RETURN_FIQ_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-END_FUNC vector_fiq_entry
-
-LOCAL_FUNC vector_cpu_on_entry , :
-	adr	x16, thread_cpu_on_handler_ptr
-	ldr	x16, [x16]
-	blr	x16
-	mov	x1, x0
-	ldr	x0, =TEESMC_OPTEED_RETURN_ON_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-END_FUNC vector_cpu_on_entry
-
-LOCAL_FUNC vector_cpu_off_entry , :
-	adr	x16, thread_cpu_off_handler_ptr
-	ldr	x16, [x16]
-	blr	x16
-	mov	x1, x0
-	ldr	x0, =TEESMC_OPTEED_RETURN_OFF_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-END_FUNC vector_cpu_off_entry
-
-LOCAL_FUNC vector_cpu_suspend_entry , :
-	adr	x16, thread_cpu_suspend_handler_ptr
-	ldr	x16, [x16]
-	blr	x16
-	mov	x1, x0
-	ldr	x0, =TEESMC_OPTEED_RETURN_SUSPEND_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-END_FUNC vector_cpu_suspend_entry
-
-LOCAL_FUNC vector_cpu_resume_entry , :
-	adr	x16, thread_cpu_resume_handler_ptr
-	ldr	x16, [x16]
-	blr	x16
-	mov	x1, x0
-	ldr	x0, =TEESMC_OPTEED_RETURN_RESUME_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-END_FUNC vector_cpu_resume_entry
-
-LOCAL_FUNC vector_system_off_entry , :
-	adr	x16, thread_system_off_handler_ptr
-	ldr	x16, [x16]
-	blr	x16
-	mov	x1, x0
-	ldr	x0, =TEESMC_OPTEED_RETURN_SYSTEM_OFF_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-END_FUNC vector_system_off_entry
-
-LOCAL_FUNC vector_system_reset_entry , :
-	adr	x16, thread_system_reset_handler_ptr
-	ldr	x16, [x16]
-	blr	x16
-	mov	x1, x0
-	ldr	x0, =TEESMC_OPTEED_RETURN_SYSTEM_RESET_DONE
-	smc	#0
-	b	.	/* SMC should not return */
-END_FUNC vector_system_reset_entry
-
-/*
- * Vector table supplied to ARM Trusted Firmware (ARM-TF) at
- * initialization.
- *
- * Note that ARM-TF depends on the layout of this vector table, any change
- * in layout has to be synced with ARM-TF.
- */
-FUNC thread_vector_table , :
-	b	vector_std_smc_entry
-	b	vector_fast_smc_entry
-	b	vector_cpu_on_entry
-	b	vector_cpu_off_entry
-	b	vector_cpu_resume_entry
-	b	vector_cpu_suspend_entry
-	b	vector_fiq_entry
-	b	vector_system_off_entry
-	b	vector_system_reset_entry
-END_FUNC thread_vector_table
-KEEP_PAGER thread_vector_table
-
 
 /* void thread_resume(struct thread_ctx_regs *regs) */
 FUNC thread_resume , :
@@ -175,93 +50,6 @@ FUNC thread_resume , :
 	store_xregs sp, THREAD_CORE_LOCAL_X0, 0, 1
 	b	eret_to_el0
 END_FUNC thread_resume
-
-FUNC thread_std_smc_entry , :
-	/* pass x0-x7 in a struct thread_smc_args */
-	sub	sp, sp, #THREAD_SMC_ARGS_SIZE
-	store_xregs sp, THREAD_SMC_ARGS_X0, 0, 7
-	mov	x0, sp
-
-	/* Call the registered handler */
-	bl	__thread_std_smc_entry
-
-	/*
-	 * Load the returned x0-x3 into preserved registers and skip the
-	 * "returned" x4-x7 since they will not be returned to normal
-	 * world.
-	 */
-	load_xregs sp, THREAD_SMC_ARGS_X0, 20, 23
-	add	sp, sp, #THREAD_SMC_ARGS_SIZE
-
-	/* Mask all maskable exceptions before switching to temporary stack */
-	msr	daifset, #DAIFBIT_ALL
-	bl	thread_get_tmp_sp
-	mov	sp, x0
-
-	bl	thread_state_free
-
-	ldr	x0, =TEESMC_OPTEED_RETURN_CALL_DONE
-	mov	x1, x20
-	mov	x2, x21
-	mov	x3, x22
-	mov	x4, x23
-	smc	#0
-	b	.	/* SMC should not return */
-END_FUNC thread_std_smc_entry
-
-/* void thread_rpc(uint32_t rv[THREAD_RPC_NUM_ARGS]) */
-FUNC thread_rpc , :
-	/* Read daif and create an SPSR */
-	mrs	x1, daif
-	orr	x1, x1, #(SPSR_64_MODE_EL1 << SPSR_64_MODE_EL_SHIFT)
-
-	/* Mask all maskable exceptions before switching to temporary stack */
-	msr	daifset, #DAIFBIT_ALL
-	push	x0, xzr
-	push	x1, x30
-	bl	thread_get_ctx_regs
-	ldr	x30, [sp, #8]
-	store_xregs x0, THREAD_CTX_REGS_X19, 19, 30
-	mov	x19, x0
-
-	bl	thread_get_tmp_sp
-	pop	x1, xzr		/* Match "push x1, x30" above */
-	mov	x2, sp
-	str	x2, [x19, #THREAD_CTX_REGS_SP]
-	ldr	x20, [sp]	/* Get pointer to rv[] */
-	mov	sp, x0		/* Switch to tmp stack */
-	/*
-	 * We need to read rv[] early, because thread_state_suspend
-	 * can invoke virt_unset_guest() which will unmap pages,
-	 * where rv[] resides
-	 */
-	load_wregs x20, 0, 21, 23	/* Load rv[] into w20-w22 */
-
-	adr	x2, .thread_rpc_return
-	mov	w0, #THREAD_FLAGS_COPY_ARGS_ON_RETURN
-	bl	thread_state_suspend
-	mov	x4, x0		/* Supply thread index */
-	ldr	w0, =TEESMC_OPTEED_RETURN_CALL_DONE
-	mov	x1, x21
-	mov	x2, x22
-	mov	x3, x23
-	smc	#0
-	b	.		/* SMC should not return */
-
-.thread_rpc_return:
-	/*
-	 * At this point has the stack pointer been restored to the value
-	 * stored in THREAD_CTX above.
-	 *
-	 * Jumps here from thread_resume above when RPC has returned. The
-	 * IRQ and FIQ bits are restored to what they where when this
-	 * function was originally entered.
-	 */
-	pop	x16, xzr	/* Get pointer to rv[] */
-	store_wregs x16, 0, 0, 5	/* Store w0-w5 into rv[] */
-	ret
-END_FUNC thread_rpc
-KEEP_PAGER thread_rpc
 
 FUNC thread_smc , :
 	smc	#0
@@ -1090,14 +878,13 @@ END_FUNC el0_sync_abort
 	mrs	x1, spsr_el1
 	mrs	x2, elr_el1
 	bl	thread_state_suspend
-	mov	w4, w0		/* Supply thread index */
 
 	/* Update core local flags */
 	/* Switch to SP_EL1 */
 	msr	spsel, #1
-	ldr	w0, [sp, #THREAD_CORE_LOCAL_FLAGS]
-	lsr	w0, w0, #THREAD_CLF_SAVED_SHIFT
-	str	w0, [sp, #THREAD_CORE_LOCAL_FLAGS]
+	ldr	w1, [sp, #THREAD_CORE_LOCAL_FLAGS]
+	lsr	w1, w1, #THREAD_CLF_SAVED_SHIFT
+	str	w1, [sp, #THREAD_CORE_LOCAL_FLAGS]
 	msr	spsel, #0
 
 	/*
@@ -1106,13 +893,8 @@ END_FUNC el0_sync_abort
 	 * set.
 	 */
 
-	ldr	w0, =TEESMC_OPTEED_RETURN_CALL_DONE
-	ldr	w1, =OPTEE_SMC_RETURN_RPC_FOREIGN_INTR
-	mov	w2, #0
-	mov	w3, #0
-	/* w4 is already filled in above */
-	smc	#0
-	b	.	/* SMC should not return */
+	/* Passing thread index in w0 */
+	b	thread_foreign_intr_exit
 .endm
 
 /*
